+++ src/kernel/dev/logical/fs/vfs/vfs.c
@@ -95,7 +95,7 @@ void vfs_close(struct filesystem_node* fs_node) {
     PANIC("not implemented");
 }
 
-struct filesystem_node* vfs_find_node_by_id(struct filesystem_node* fs_node, uint32_t id) {
+struct filesystem_node* vfs_find_node_by_id(struct filesystem_node* fs_node, uint64_t id) {
     ASSERT_NOT_NULL(fs_node);
     ASSERT_NOT_NULL(fs_node->filesystem_device);
     ASSERT_NOT_NULL(fs_node->filesystem_device->device_data);
