+++ src/kernel/sys/deviceapi/deviceapi_filesystem.h
@@ -63,7 +63,7 @@ typedef void (*filesystem_close_function)(struct filesystem_node* fs_node);
 /*
 * find a node by id
 */
-typedef struct filesystem_node* (*filesystem_find_node_by_id_function)(struct filesystem_node* fs_node, uint32_t id);
+typedef struct filesystem_node* (*filesystem_find_node_by_id_function)(struct filesystem_node* fs_node, uint64_t id);
 /*
 * get directory list.  fills struct. 
 */
